requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
plotly>=5.17.0
pandas>=2.0.0
Pillow>=10.0.0
scikit-learn>=1.3.0
groq>=0.4.0
fpdf2>=2.7.0
